refactor: 주문 처리 필터링 컴포넌트 분리 1. Filter 컴포넌트로 분리
